rename cards into magaras
I wasn't convinced by the card-flavour, a little too random and wasn't
relying on deck-building, it wasn't suited for character
progression/diferentiation. The new magaras are a little like old
rods, but without the charges (I think it will stay like this). You will
have a limited number of slots for rods, and will be able to chose which
ones to keep (TODO).